+++ Realization/frontend/czechidm-acc/src/content/system/SystemMappingDetail.js
@@ -353,10 +353,10 @@ class SystemMappingDetail extends Advanced.AbstractTableContent {
                 uiKey={uiKeyAttributes}
                 manager={systemAttributeMappingManager}
                 forceSearchParameters={forceSearchParameters}
-                showRowSelection={Managers.SecurityManager.hasAnyAuthority(['SYSTEM_UPDATE'])}
+                showRowSelection={Managers.SecurityManager.hasAnyAuthority(['SYSTEMATTRIBUTEMAPPING_UPDATE'])}
                 rowClass={({rowIndex, data}) => { return data[rowIndex].disabledAttribute ? 'disabled' : ''; }}
                 actions={
-                  Managers.SecurityManager.hasAnyAuthority(['SYSTEM_UPDATE'])
+                  Managers.SecurityManager.hasAnyAuthority(['SYSTEMATTRIBUTEMAPPING_DELETE'])
                   ?
                   [{ value: 'delete', niceLabel: this.i18n('action.delete.action'), action: this.onDelete.bind(this), disabled: false }]
                   :
@@ -369,7 +369,7 @@ class SystemMappingDetail extends Advanced.AbstractTableContent {
                       key="add_button"
                       className="btn-xs"
                       onClick={this.showDetail.bind(this, { }, true)}
-                      rendered={Managers.SecurityManager.hasAnyAuthority(['SYSTEM_UPDATE'])}>
+                      rendered={Managers.SecurityManager.hasAnyAuthority(['SYSTEMATTRIBUTEMAPPING_UPDATE'])}>
                       <Basic.Icon type="fa" icon="plus"/>
                       {' '}
                       {this.i18n('button.add')}